+++ include/iparticlestage.h
@@ -2,9 +2,9 @@
 #define _IPARTICLE_STAGE_H_
 
 template<typename Element> class BasicVector3;
-typedef BasicVector3<double> Vector3;
+typedef BasicVector3<float> Vector3;
 template<typename Element> class BasicVector4;
-typedef BasicVector4<double> Vector4;
+typedef BasicVector4<float> Vector4;
 
 namespace particles
 {
