+++ plugins/script/interfaces/MathInterface.cpp
@@ -12,26 +12,26 @@ namespace script {
 // IScriptInterface implementation
 void MathInterface::registerInterface(boost::python::object& nspace) {
 	// Add the Vector3 class
-	nspace["Vector3"] = boost::python::class_<Vector3>("Vector3", boost::python::init<double, double, double>())
+	nspace["Vector3"] = boost::python::class_<Vector3>("Vector3", boost::python::init<float, float, float>())
 		.def(boost::python::init<const Vector3&>())
 		// greebo: Pick the correct overload - this is hard to read, but it is necessary
-		.def("x", static_cast<double& (Vector3::*)()>(&Vector3::x),
+		.def("x", static_cast<float& (Vector3::*)()>(&Vector3::x),
 			boost::python::return_value_policy<boost::python::copy_non_const_reference>())
-		.def("y", static_cast<double& (Vector3::*)()>(&Vector3::y),
+		.def("y", static_cast<float& (Vector3::*)()>(&Vector3::y),
 			boost::python::return_value_policy<boost::python::copy_non_const_reference>())
-		.def("z", static_cast<double& (Vector3::*)()>(&Vector3::z),
+		.def("z", static_cast<float& (Vector3::*)()>(&Vector3::z),
 			boost::python::return_value_policy<boost::python::copy_non_const_reference>())
 		.def("getLength", &Vector3::getLength)
 		.def("getLengthSquared", &Vector3::getLengthSquared)
 		.def("getNormalised", &Vector3::getNormalised)
 		.def("normalise", &Vector3::normalise)
 		.def("getInversed", &Vector3::getInversed)
-		.def("dot", &Vector3::dot<double>)
-		.def("angle", &Vector3::angle<double>)
-		.def("crossProduct", &Vector3::crossProduct<double>)
+		.def("dot", &Vector3::dot<float>)
+		.def("angle", &Vector3::angle<float>)
+		.def("crossProduct", &Vector3::crossProduct<float>)
 		.def("max", &Vector3::max)
 		.def("min", &Vector3::min)
-		.def("isParallel", &Vector3::isParallel<double>)
+		.def("isParallel", &Vector3::isParallel<float>)
 		// Most important operators
 		.def(boost::python::self + boost::python::self)		// __add__
 		.def(boost::python::self - boost::python::self)		// __sub__
@@ -44,21 +44,21 @@ void MathInterface::registerInterface(boost::python::object& nspace) {
 	nspace["Vertex3f"] = boost::python::class_<Vertex3f,
 		boost::python::bases<Vector3> >("Vertex3f", boost::python::init<>() )
 		.def(boost::python::init<const Vector3&>())
-		.def(boost::python::init<double, double, double>())
+		.def(boost::python::init<float, float, float>())
 	;
 
 	// Add the Vector2 class
-	nspace["Vector2"] = boost::python::class_<Vector2>("Vector2", boost::python::init<double, double>())
+	nspace["Vector2"] = boost::python::class_<Vector2>("Vector2", boost::python::init<float, float>())
 		.def(boost::python::init<const Vector2&>())
 		// greebo: Pick the correct overload - this is hard to read, but it is necessary
-		.def("x", static_cast<double& (Vector2::*)()>(&Vector2::x),
+		.def("x", static_cast<float& (Vector2::*)()>(&Vector2::x),
 			boost::python::return_value_policy<boost::python::copy_non_const_reference>())
-		.def("y", static_cast<double& (Vector2::*)()>(&Vector2::y),
+		.def("y", static_cast<float& (Vector2::*)()>(&Vector2::y),
 			boost::python::return_value_policy<boost::python::copy_non_const_reference>())
 		.def("getLength", &Vector2::getLength)
 		.def("getLengthSquared", &Vector2::getLengthSquared)
-		.def("dot", &Vector2::dot<double>)
-		.def("crossProduct", &Vector2::crossProduct<double>)
+		.def("dot", &Vector2::dot<float>)
+		.def("crossProduct", &Vector2::crossProduct<float>)
 		// Most important operators
 		.def(boost::python::self + boost::python::self)		// __add__
 		.def(boost::python::self - boost::python::self)		// __sub__
@@ -68,22 +68,22 @@ void MathInterface::registerInterface(boost::python::object& nspace) {
 	;
 
 	// Add the Vector4 class
-	boost::python::class_<Vector4> vector4Decl("Vector4", boost::python::init<double, double, double, double>());
+	boost::python::class_<Vector4> vector4Decl("Vector4", boost::python::init<float, float, float, float>());
 
 	vector4Decl.def(boost::python::init<const Vector4&>())
 		// greebo: Pick the correct overload - this is hard to read, but it is necessary
-		.def("x", static_cast<double& (Vector4::*)()>(&Vector4::x),
+		.def("x", static_cast<float& (Vector4::*)()>(&Vector4::x),
 			boost::python::return_value_policy<boost::python::copy_non_const_reference>())
-		.def("y", static_cast<double& (Vector4::*)()>(&Vector4::y),
+		.def("y", static_cast<float& (Vector4::*)()>(&Vector4::y),
 			boost::python::return_value_policy<boost::python::copy_non_const_reference>())
-		.def("z", static_cast<double& (Vector4::*)()>(&Vector4::z),
+		.def("z", static_cast<float& (Vector4::*)()>(&Vector4::z),
 			boost::python::return_value_policy<boost::python::copy_non_const_reference>())
-		.def("w", static_cast<double& (Vector4::*)()>(&Vector4::w),
+		.def("w", static_cast<float& (Vector4::*)()>(&Vector4::w),
 			boost::python::return_value_policy<boost::python::copy_non_const_reference>())
 		.def("getVector3", static_cast<Vector3& (Vector4::*)()>(&Vector4::getVector3),
 			boost::python::return_value_policy<boost::python::copy_non_const_reference>())
 		.def("getProjected", &Vector4::getProjected)
-		.def("dot", &Vector4::dot<double>)
+		.def("dot", &Vector4::dot<float>)
 		// Most important operators
 		.def(boost::python::self + boost::python::self)		// __add__
 		.def(boost::python::self - boost::python::self)		// __sub__
